+++ docs/assets/circle-DxUgLc9F.js
@@ -1,0 +1,1 @@
+import{gt as e}from"./index-vEWl5V_M.js";var t=e(`circle`,[[`circle`,{cx:`12`,cy:`12`,r:`10`,key:`1mglay`}]]);export{t};